+++ client/src/components/application/bookmark/vendor-coupon.jsx
@@ -4,7 +4,7 @@ import { Table, Container, Row, Col, Card, CardBody, CardHeader, Nav, NavItem, T
 import {toast} from 'react-toastify';
 import axios from 'axios'
 import { useHistory } from 'react-router-dom'
-
+import SweetAlert from 'sweetalert2'
 
 const VendorCoupon = (props) => {
 
@@ -33,20 +33,42 @@ const VendorCoupon = (props) => {
 
        // Delete functionality
 
-   const handleDelete = (id) => {
-    const config = {
-      headers: { 'Content-Type': 'application/json'  ,'Access-Control-Allow-Origin': '*' , 'Authorization': 'JWT '+token }
-      };
-
-      axios.delete(`/api/vendor-coupons/`+`${id}`,config
-      ) .then(response => {
-        toast.success("Coupon Deleted !")
-        setTimeout(() => {
-          window.location.reload();
-        }, 1000);
-      })
-         .catch(error => console.log('Form submit error', error))
+  const handleDelete = (id) => {
+    SweetAlert.fire({
+      title: 'Are you sure?',
+      text: "Once deleted, you will not be able to recover this coupon!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Ok',
+      cancelButtonText: 'cancel',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.value) {
+        const config = {
+          headers: { 'Content-Type': 'application/json'  ,'Access-Control-Allow-Origin': '*' , 'Authorization': 'JWT '+token }
+          };
     
+          axios.delete(`/api/vendor-coupons/`+`${id}`,config
+          ) .then(response => {
+            toast.success("Coupon Deleted !")
+            setTimeout(() => {
+              window.location.reload();
+            }, 1000);
+          })
+             .catch(error => console.log('Form submit error', error))
+        
+        SweetAlert.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+      }
+      else {
+        SweetAlert.fire(
+          'Coupon is safe!'
+        )
+      }
+    })
   }
