+++ client/src/components/application/bookmark/vendor-photogallery.jsx
@@ -5,6 +5,7 @@ import { Container, Row, Col,FormGroup , Button , Card, CardHeader, CardBody } f
 import {toast} from 'react-toastify';
 import axios from 'axios'
 import { useHistory } from 'react-router-dom'
+import SweetAlert from 'sweetalert2'
 
 const VendorPhotogallery = () => {
 
@@ -43,20 +44,43 @@ const VendorPhotogallery = () => {
 
    // Delete functionality
 
-   const handleDelete = (id) => {
-    const config = {
-      headers: { 'Content-Type': 'application/json'  ,'Access-Control-Allow-Origin': '*' , 'Authorization': 'JWT '+token }
-      };
+  const handleDelete = (id) => {
+    SweetAlert.fire({
+      title: 'Are you sure?',
+      text: "Once deleted, you will not be able to recover this coupon!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Ok',
+      cancelButtonText: 'cancel',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.value) {
+          
+        const config = {
+            headers: { 'Content-Type': 'application/json'  ,'Access-Control-Allow-Origin': '*' , 'Authorization': 'JWT '+token }
+            };
+      
+            axios.delete(`/api/gallery/`+`${id}`,config
+            ) .then(response => {
+              toast.success("Image Deleted !")
+              setTimeout(() => {
+                  window.location.reload();
+              }, 1000);
+            })
+               .catch(error => console.log('Form submit error', error))
 
-      axios.delete(`/api/gallery/`+`${id}`,config
-      ) .then(response => {
-        toast.success("Image Deleted !")
-        setTimeout(() => {
-            window.location.reload();
-        }, 1000);
-      })
-         .catch(error => console.log('Form submit error', error))
-    
+        SweetAlert.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+      }
+      else {
+        SweetAlert.fire(
+          'Coupon is safe!'
+        )
+      }
+    })
   }
 
   const handleSubmit = event => {
